chore: move methods in layout into config file
